+++ nucypher-core/src/reencryption.rs
@@ -27,6 +27,10 @@ pub struct ReencryptionRequest {
     pub publisher_verifying_key: PublicKey,
     /// Recipient's (Bob's) verifying key.
     pub bob_verifying_key: PublicKey,
+    /// A blob of bytes containing decryption conditions for this message.
+    pub conditions: Option<Box<[u8]>>,
+    /// A blob of bytes containing context required to evaluate conditions.
+    pub context: Option<Box<[u8]>>,
 }
 
 impl ReencryptionRequest {
@@ -37,6 +41,8 @@ impl ReencryptionRequest {
         encrypted_kfrag: &EncryptedKeyFrag,
         publisher_verifying_key: &PublicKey,
         bob_verifying_key: &PublicKey,
+        conditions: Option<&[u8]>,
+        context: Option<&[u8]>,
     ) -> Self {
         Self {
             capsules: capsules.into(),
@@ -44,6 +50,8 @@ impl ReencryptionRequest {
             encrypted_kfrag: encrypted_kfrag.clone(),
             publisher_verifying_key: *publisher_verifying_key,
             bob_verifying_key: *bob_verifying_key,
+            conditions: conditions.map(|c| c.into()),
+            context: context.map(|c| c.into()),
         }
     }
 }
@@ -182,3 +190,56 @@ impl<'a> ProtocolObjectInner<'a> for ReencryptionResponse {
 }
 
 impl<'a> ProtocolObject<'a> for ReencryptionResponse {}
+
+#[cfg(test)]
+mod tests {
+    use umbral_pre::SecretKey;
+    use umbral_pre::{encrypt, generate_kfrags, Signer};
+
+    use crate::{EncryptedKeyFrag, HRAC};
+
+    use super::ReencryptionRequest;
+
+    #[test]
+    fn conditions_and_context_are_different() {
+        let some_secret = SecretKey::random();
+        let some_trinket = some_secret.public_key();
+
+        let _another_secret = SecretKey::random();
+        let another_trinket = some_secret.public_key();
+
+        let encryption_result = encrypt(&some_trinket, b"peace at dawn");
+
+        let (capsule, _ciphertext) = encryption_result.unwrap();
+
+        let hrac = HRAC::new(&some_trinket, &another_trinket, &[42]);
+
+        let signer = Signer::new(SecretKey::random());
+
+        let verified_kfrags =
+            generate_kfrags(&some_secret, &another_trinket, &signer, 5, 8, true, true);
+        let verified_kfrags_vector = verified_kfrags.into_vec();
+        let one_verified_krag_in_particular = verified_kfrags_vector[0].clone();
+        let encrypted_kfrag = EncryptedKeyFrag::new(
+            &signer,
+            &another_trinket,
+            &hrac,
+            one_verified_krag_in_particular,
+        );
+
+        let request = ReencryptionRequest::new(
+            &[capsule],
+            &hrac,
+            &encrypted_kfrag,
+            &some_trinket,
+            &another_trinket,
+            Some(&[47u8]),
+            Some(&[51u8]),
+        );
+        let conditions = request.conditions.unwrap();
+        assert!(conditions[0].eq(&47u8));
+
+        let context = request.context.unwrap();
+        assert!(context[0].eq(&51u8));
+    }
+}